Clean up package structure.
