--- src/main/java/com/github/corese4rch/cvurl/easytest/domain/asserts/property/Url.java
@@ -1,16 +1,0 @@
-package com.github.corese4rch.cvurl.easytest.domain.asserts.property;
-
-import com.github.corese4rch.cvurl.easytest.domain.EasyRequest;
-
-public class Url extends RequestProperty<String> {
-
-    @Override
-    protected String getName() {
-        return "url";
-    }
-
-    @Override
-    protected String mapProperty(EasyRequest request) {
-        return request.getUrl();
-    }
-}
