--- src/main/java/com/github/corese4rch/cvurl/easytest/domain/asserts/rule/Rule.java
@@ -1,11 +1,0 @@
-package com.github.corese4rch.cvurl.easytest.domain.asserts.rule;
-
-public abstract class Rule<T> {
-    public abstract boolean isValid(T obj);
-
-    public abstract String getDescription();
-
-    public final boolean isNotValid(T obj) {
-        return !isValid(obj);
-    }
-}
