--- src/test/java/com/github/corese4rch/cvurl/easytest/domain/utils/Urls.java
@@ -1,11 +1,0 @@
-package com.github.corese4rch.cvurl.easytest.domain.utils;
-
-import static java.lang.String.format;
-
-public class Urls {
-    public static final String TEST_URL = "http://test.com";
-
-    private Urls() {
-        throw new IllegalStateException(format("Creation of class %s is forbidden", Urls.class));
-    }
-}
